+++ internal/config/config_test.go
@@ -1,7 +1,0 @@
-// Copyright (c) 2025 Rich Haase
-// Licensed under the MIT License. See LICENSE file in the project root for license information.
-
-package config
-
-// TOML configuration tests removed - project now uses YAML configuration exclusively.
-// See yaml_config_test.go for current configuration testing.